+++ example/example-types.dart
@@ -1,22 +1,22 @@
 import 'package:vnum/vnum.dart';
 
-@EnumReflectable
-class CarType extends VNum<String> {
+@VnumDefinition
+class CarType extends Vnum<String> {
   static final CarType sedan = CarType.define("sedan-value");
   static final CarType suv = CarType.define("suv-value");
   static final CarType truck = CarType.define("truck-value");
   static final CarType none = CarType.define("value-none");
   CarType.define(String fromValue) : super.define(fromValue);
-  factory CarType(String value) => VNum.fromValue(value,CarType);
+  factory CarType(String value) => Vnum.fromValue(value,CarType);
 }
 
-@EnumReflectable
-class Fruit extends VNum<int> {
+@VnumDefinition
+class Fruit extends Vnum<int> {
   static final Fruit apple = Fruit.define(1);
   static final Fruit orange = Fruit.define(2);
   static final Fruit banana = Fruit.define(3);
   Fruit.define(int fromValue) : super.define(fromValue);
-  factory Fruit(int value) => VNum.fromValue(value,Fruit);
+  factory Fruit(int value) => Vnum.fromValue(value,Fruit);
 
   String color(){
     if (value == Fruit.apple.value) {