+++ test/test_types.dart
@@ -1,11 +1,11 @@
 import 'package:vnum/vnum.dart';
 
-@EnumReflectable
-class CarType extends VNum<String> {
+@VnumDefinition
+class CarType extends Vnum<String> {
   static final CarType sedan = CarType.define("sedan-value");
   static final CarType suv = CarType.define("suv-value");
   static final CarType truck = CarType.define("truck-value");
   static final CarType none = CarType.define("value-none");
   CarType.define(String fromValue) : super.define(fromValue);
-  factory CarType(String value) => VNum.fromValue(value,CarType);
+  factory CarType(String value) => Vnum.fromValue(value,CarType);
 }