+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [1.0.2] - Unify keywords
+ * ```VNum``` has been renamed to ```Vnum```
+ * ```@EnumReflectable``` has been renamed to ```@VnumDefinition```
+ * Update documentation
 ## [1.0.1] - Update documentation
  * Update documentation
  * Add example folder